Refactor Azure client for future extendability (#12)
This change refactors Azure client so that the `azure.Client` acts as a
kind of facade for all different Azure service clients needed. Each
service client is then an own member of `azure.Client` and implements an
interface under `azure` package.

The concrete implementation of a service specific client is a wrapped
version of actual Azure SDK client which allows for easy implementation
of higher level functionality while reducing boilerplate needed to call
for actual SDK functions.

The involved types are also wrapped by `azure` package directly from SDK
so that the concrete version of SDK API can be hidden and easily
updated, but the type still provides full access to SDK fields.